requirements.txt
requests
mysql-connector-python
pandas
streamlit
plotly
openpyxl
scikit-learn
streamlit-autorefresh